+++ src/main/java/hr/dlatecki/algorithms/gen_alg/codecs/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Package which contains interfaces which define codecs.
+ * 
+ * @author Domagoj Latečki
+ * @since 1.8
+ */
+package hr.dlatecki.algorithms.gen_alg.codecs;