+++ src/main/java/hr/dlatecki/algorithms/gen_alg/operators/abstracts/AbstractCrossoverOperator.java
@@ -11,7 +11,7 @@ import hr.dlatecki.algorithms.gen_alg.population.interfaces.IChromosome;
  * @author Domagoj Latečki
  * @version 1.0
  * @since 1.8
- * @param <C> Type of chromosome which will be used in the crossover operator.
+ * @param <C> the type of chromosome which will be used in the crossover operator.
  * @see IChromosome
  * @see ICrossoverOperator
  */