+++ src/main/java/hr/dlatecki/algorithms/gen_alg/operators/abstracts/AbstractSelectionOperator.java
@@ -12,7 +12,7 @@ import hr.dlatecki.algorithms.gen_alg.population.interfaces.IChromosome;
  * @author Domagoj Latečki
  * @version 1.1
  * @since 1.8
- * @param <C> Type of chromosome which will be used in the selection operator.
+ * @param <C> the type of chromosome which will be used in the selection operator.
  * @see IChromosome
  * @see ISelectionOperator
  */