+++ src/main/java/hr/dlatecki/algorithms/gen_alg/population/abstracts/AbstractPopulationGenerator.java
@@ -12,7 +12,7 @@ import hr.dlatecki.algorithms.gen_alg.population.interfaces.IPopulationGenerator
  * @author Domagoj Latečki
  * @version 2.0
  * @since 1.8
- * @param <C> Type of chromosome which will be used in the population generator.
+ * @param <C> the type of chromosome which will be used in the population generator.
  * @see IChromosome
  * @see IPopulationGenerator
  */