+++ src/main/java/hr/dlatecki/algorithms/gen_alg/population/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * Package which contains implementations of the chromosomes and population generators used in genetic algorithms
+ * defined in {@link hr.dlatecki.algorithms.gen_alg.impl} package. Also contains sub-packages which define interfaces
+ * and abstract classes for chromosomes, population generators and fitness tests.
+ * 
+ * @author Domagoj Latečki
+ * @since 1.8
+ */
+package hr.dlatecki.algorithms.gen_alg.population;